move login session buttons to layout
